Finally get how to convert strings!!
So it seems that `std.mem.trim` (and likely other functions) don't work
correctly on C strings, or don't work correctly if passed on and coerced
weirdly.

But what helps is converting to a regular zig string/slice using
`std.mem.sliceTo(c_string, 0)`.  It even makes sense, given that `0` is
the nul byte at the end of a c string.  Huh.

Oh, and we run Go thingy now!!