Actually avoid rerendering if no new output or inputs  🚅
Previously we would not skip rerendering because the `continue`
statement only skipped in the for-loop through the commands, not the
actual render loop.

We now do that and also re-render (but do not restart the command) if
there is a navigation or other input.

And finally the `SDL_Delay` call is `defer`ed now so that it actually
executes with every loop iteration, including the `continue` when we're
not rerendering.  Otherwise we'd use even more cpu because we'd have an
always-busy loop.

Fun!